requirements.txt
python-dotenv~=1.0.1
<<<<<<< HEAD
requests~=2.32.3
=======
requests~=2.32.3
>>>>>>> de7a2db2c11e85cb4978d18515d9f267fcd3c52b
